requirements.txt
django==2.2.6
pillow==8.4.0
djangorestframework==3.11.0
djoser
drf-extra-fields==3.2.1
django-filter==21.1